backend/requirements.txt
Django==5.1.4
djangorestframework==3.15.2
django-cors-headers==4.6.0
django-filter==24.3
django-environ==0.11.2
django-redis==5.4.0
django-celery-beat==2.7.0
django-celery-results==2.5.1
django-extensions==3.2.3

# Database
psycopg2-binary==2.9.10
dj-database-url==2.3.0

# Authentication
djangorestframework-simplejwt==5.4.0
django-rest-framework-api-key==3.0.0

# Async / WebSocket
channels==4.2.0
channels-redis==4.2.1
daphne==4.1.2

# Task Queue
celery[redis]==5.4.0
redis==5.2.1

# Template rendering
Jinja2==3.1.4
MarkupSafe==3.0.2

# Email providers
sendgrid==6.11.0

# SMS providers
twilio==9.4.0

# Push notifications
firebase-admin==6.6.0

# Slack
slack-sdk==3.33.5

# Telegram
python-telegram-bot==21.8

# HTTP
requests==2.32.3
httpx==0.28.1

# Utilities
python-dateutil==2.9.0
pytz==2024.2
uuid6==2024.7.10

# Monitoring
sentry-sdk==2.19.2

# Code quality
flake8==7.1.1
black==24.10.0
isort==5.13.2

# Testing
pytest==8.3.4
pytest-django==4.9.0
pytest-cov==6.0.0
factory-boy==3.3.1

# API Documentation
drf-spectacular==0.28.0

# Gunicorn (production WSGI - fallback)
gunicorn==23.0.0
